+++ climpred/bootstrap.py
@@ -420,6 +420,7 @@ def bootstrap_compute(
             verif,
             metric=metric,
             comparison=comparison,
+            alignment=alignment,
             add_attrs=False,
             dim=dim,
             **metric_kwargs,
@@ -452,7 +453,14 @@ def bootstrap_compute(
         # impossible for probabilistic
         if not metric.probabilistic:
             pers.append(
-                reference_compute(smp_hind, verif, metric=metric, **metric_kwargs)
+                reference_compute(
+                    smp_hind,
+                    verif,
+                    metric=metric,
+                    alignment=alignment,
+                    add_attrs=False,
+                    **metric_kwargs,
+                )
             )
     init = xr.concat(init, dim='bootstrap', **CONCAT_KWARGS)
     uninit = xr.concat(uninit, dim='bootstrap', **CONCAT_KWARGS)